+++ .buildkite/pipelines/pull-request/flakiness-detection.yml
@@ -1,9 +1,9 @@
 steps:
-  - group: "repeat-changed-tests"
+  - group: "flakiness-detection"
     steps:
       - label: "detect changed tests"
         command: |
           npm install -g bun@1.0.4
-          bun .buildkite/scripts/repeat-changed-tests.ts
+          bun .buildkite/scripts/flakiness-detection/index.ts
         timeout_in_minutes: 10
         soft_fail: true
